tests/requirements.txt
pytest
requests
python-dotenv
pytest-xdist
pytest-rerunfailures